Fix intermittent failure in Android form submission test
